converted long into int

+++ NumberOfYears.java
@@ -9,8 +9,8 @@ public class NumberOfYears {
         System.out.print("Enter minutes: ");
         long minutes = input.nextLong();
 
-        long years = minutes / (60 * 24 * 365);
-        long days = (minutes % (60 * 24 * 365)) / (60 * 24);
+        int years = (int)(minutes / (60 * 24 * 365));
+        int days = (int)((minutes % (60 * 24 * 365)) / (60 * 24));
 
         System.out.println("It is " + years + " year(s) and " + days + " day(s).");
     }
